+++ CHANGELOG.md
@@ -13,7 +13,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - mongo_libs.create_cmd: Changed cmds_gen.add_cmd to gen_libs.add_cmd and cmds_gen.is_add_cmd to gen_libs.is_add_cmd.
 
 ### Deprecated
-- mongo_libs.create_slv_array:  Not required.
+- mongo_libs.create_slv_array
 
 ### Removed
 - mongo_class.RepSetColl.\_db_auth method